Refactor libinput and input
