+++ meson.build
@@ -2,7 +2,7 @@ project(
   'z11',
   'c', 'cpp',
   license : 'Apache-2.0',
-  default_options : ['warning_level=3', 'werror=true', 'cpp_std=c++17', 'optimization=2'],
+  default_options : ['warning_level=3', 'werror=true', 'cpp_std=c++17', 'optimization=2', 'ggdb=true'],
   meson_version : '>=0.57.0',
 )
