+++ zazen/z_server.h
@@ -10,6 +10,7 @@ class ZServer
   bool Init();
   void Poll();
   void Frame();
+  void Shutdown();
 
   class RenderStateIterator
   {
